+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.3.0] - 2026-06-02
 
 - Add Rails 8.0 & 8.1 as known versions
+- Sort functions and triggers
 
 ## [0.2.0] - 2024-12-03